requirements.txt
## The following requirements were added by pip freeze:
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
black==25.1.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
django-stubs==5.2.2
django-stubs-ext==5.2.2
filetype==1.2.0
frozenlist==1.7.0
langchain==1.0.2
redis==5.0.1
ipython==8.19.0
tuspy
huey
google-genai==1.46.0
langchain-google-genai==3.0.0
psycopg2-binary==2.9.10
boto3==1.34.144
django-storages
gunicorn